requirements.txt
beautifulsoup4==4.11.2
discord.py==1.7.3
requests==2.27.1
lxml==3.7.3